rename : ResultRacingRenderer -> ResultRacingDomain
